pyproject.toml
[project]
name = "agent-tutorial"
version = "0.1.0"
description = "Simple tutorial showcasing Strands and AgentCore"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.26",
    "aws-cdk-aws-bedrock-agentcore-alpha>=2.249.0a0",
    "aws-cdk-aws-lambda-python-alpha>=2.250.0a0",
    "aws-cdk-lib>=2.249.0",
    "bedrock-agentcore[a2a]>=1.6.2",
    "boto3-stubs>=1.42.89",
    "constructs>=10.6.0",
    "fastapi>=0.135.3",
    "gradio>=6.13.0",
    "opensearch-py>=3.1.0",
    "ruff>=0.15.10",
    "strands-agents>=1.35.0",
]

[tool.pytest.ini_options]
norecursedirs = ["cdk.out", ".venv", "node_modules"]
pythonpath = ["."]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B104", "B311"]

[tool.poe.tasks.build]
help = "Runs the sync process"
cmd = "uv sync"

[tool.poe.tasks.app]
help = "Runs the code"
cmd = "uv run invoke.py"

[tool.poe.tasks.test]
help = "Run the tests with coverage"
cmd = "uv run pytest -vrrP --testdox --cov agents infra tests"

[tool.poe.tasks.formatter]
help = "Formats files using the Ruff formatter"
cmd = "uv run ruff format"

[tool.poe.tasks.lint]
help = "Checks linting using Ruff"
cmd = "uv run ruff check --fix"

[tool.poe.tasks.audit]
help = "Audits dependencies for vulnerabilities"
cmd = "uv run pip-audit"

[tool.poe.tasks.scan]
help = "Scans the application code for vulnerabilities"
cmd = "uv run bandit -r app.py agents infra"

[tool.poe.tasks.typing]
help = "Pre-run checks on typing violations"
cmd = "uv run mypy --explicit-package-bases app.py agents infra"

[tool.poe.tasks.check]
help = "Run all checks required pre-deployment"
sequence = ["lint", "audit", "scan", "typing"]

[tool.poe.tasks.bootstrap]
help = "Bootstraps the AWS CDK (run once in region/account)"
cmd = "cdk bootstrap"

[tool.poe.tasks.synth]
help = "Generate the CF Stacks into the .out file."
cmd = "cdk synth"

[tool.poe.tasks.deploy]
help = "Deploy to AWS with CDK"
cmd = "cdk deploy --all"

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "mypy>=1.20.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-testdox>=3.1.0",
    "ruff>=0.15.10",
]
